Refactor: Refactor names of functions and packages
