+++ pom.xml
@@ -69,6 +69,12 @@
             <version>3.2.0</version>
         </dependency>
 
+        <!-- Spring Data JPA -->
+        <dependency>
+            <groupId>org.springframework.data</groupId>
+            <artifactId>spring-data-jpa</artifactId>
+            <version>3.3.4</version>
+        </dependency>
         <!-- MySql -->
         <dependency>
             <groupId>com.mysql</groupId>
