+++ src/main/java/com/travelport/persistence/impl/SalesDaoImpl.java
@@ -1,7 +1,7 @@
-package com.travelport.jpa.impl;
+package com.travelport.persistence.impl;
 
 import com.travelport.entities.Sale;
-import com.travelport.jpa.SalesDao;
+import com.travelport.persistence.SalesDao;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
@@ -12,7 +12,7 @@ import java.util.List;
 
 @Transactional
 @Repository
-public class SalesDaoImpl implements SalesDao {
+public class SalesDaoImpl extends AbstractEntityDaoImpl<Sale, Integer> implements SalesDao {
 
     @PersistenceContext
     private EntityManager em;
@@ -21,11 +21,6 @@ public class SalesDaoImpl implements SalesDao {
     public void save(Sale sale) {
         em.persist(sale);
         em.flush(); // Para evitar error 500, ya que antes de crear un SaleDetail necesito persistir la Sale para que inmediatamente esté disponible su id autoGenerado para que SaleDetail lo pueda usar.
-    }
-
-    @Override
-    public void update(Sale sale) {
-        em.merge(sale);
     }
 
     @Override
